requirements.txt
requests>=2.32.3,<2.33
geopandas>=1.0.1,<1.1
python-dotenv>=1.0.1,<1.1
geopy>=2.4.1,<2.5
folium>=0.18.0,<0.19
psycopg2>=2.9.9,<3.0
neo4j>=5.25.0,<5.26
networkx>=3.2.1,<3.3
geopandas>=1.0.1,<1.1
